==> 0_DataPreprocessing.ipynb <==
import gc
import warnings
import numpy as np
import pandas as pd

warnings.filterwarnings('ignore')
pd.set_option('display.max_columns', None)

%%time
train_id = pd.read_csv('./dataset/train_identity.csv')
train_tr = pd.read_csv('./dataset/train_transaction.csv')
test_id = pd.read_csv('./dataset/test_identity.csv')
test_tr = pd.read_csv('./dataset/test_transaction.csv')

train = pd.merge(train_tr, train_id, how='left', 
                 on='TransactionID', left_index=True, 
                right_index=True)

test = pd.merge(test_tr, test_id, how='left',
               on='TransactionID', left_index=True,
               right_index=True)

del train_tr,test_id,train_id,test_tr
gc.collect()

print(train.shape)
print(test.shape)

def diffcouls(train,test):
    notin = []
    for i in train.columns:
        if i not in test.columns:
            notin.append(i)
            return(notin)
print(diffcouls(train,test))

test = test.rename(columns={"id-01": "id_01", "id-02": "id_02", "id-03": "id_03", 
                            "id-06": "id_06", "id-05": "id_05", "id-04": "id_04", 
                            "id-07": "id_07", "id-08": "id_08", "id-09": "id_09", 
                            "id-10": "id_10", "id-11": "id_11", "id-12": "id_12", 
                            "id-15": "id_15", "id-14": "id_14", "id-13": "id_13", 
                            "id-16": "id_16", "id-17": "id_17", "id-18": "id_18", 
                            "id-21": "id_21", "id-20": "id_20", "id-19": "id_19", 
                            "id-22": "id_22", "id-23": "id_23", "id-24": "id_24", 
                            "id-27": "id_27", "id-26": "id_26", "id-25": "id_25", 
                            "id-28": "id_28", "id-29": "id_29", "id-30": "id_30", 
                            "id-31": "id_31", "id-32": "id_32", "id-33": "id_33", 
                            "id-34": "id_34", "id-35": "id_35", "id-36": "id_36", 
                            "id-37": "id_37", "id-38": "id_38"})

print(diffcouls(train,test))
print(train.shape)

# Making a dataframe about total missing values and percentage of a column
def getnulls(data):
    total = data.isnull().sum().sort_values(ascending=False)
    percent = (data.isnull().sum() / len(data)).sort_values(ascending=False)
    missing_data = pd.concat([total, percent],  keys=['total', 'precent'],axis=1)
    return missing_data

# Train Data Missing Values
missing_data_train = getnulls(train)
missing_data_train.head(len(train)).T

droppedcols = missing_data_train[missing_data_train['total'] > 100000].index
print(len(droppedcols),"\n")
print(list(droppedcols))

# Drop the columns
train.drop(droppedcols, axis=1, inplace=True)
test.drop(droppedcols, axis=1, inplace=True)
print(train.shape)
print(test.shape)

#### 434 to 182 columns

missing_data_train = getnulls(train)
missing_data_train.head(len(train)).T

all_data_cols = train.columns
for i in all_data_cols:
    if train[i].dtype == 'object':
        train[i] = train[i].fillna(train[i].mode()[0])

missing_data_train = getnulls(train)
missing_data_train.head(len(train)).T

def reduce_mem_usage(df, verbose=True):
    numerics = ['int16', 'int32', 'int64', 'float16', 'float32', 'float64']
    start_mem = df.memory_usage().sum() / 1024**2
    for col in df.columns:
        col_type = df[col].dtypes
        if col_type in numerics:
            c_min = df[col].min()
            c_max = df[col].max()
            if str(col_type)[:3] == 'int':
                if c_min > np.iinfo(np.int8).min and c_max < np.iinfo(np.int8).max:
                    df[col] = df[col].astype(np.int8)
                elif c_min > np.iinfo(np.int16).min and c_max < np.iinfo(np.int16).max:
                    df[col] = df[col].astype(np.int16)
                elif c_min > np.iinfo(np.int32).min and c_max < np.iinfo(np.int32).max:
                    df[col] = df[col].astype(np.int32)
                elif c_min > np.iinfo(np.int64).min and c_max < np.iinfo(np.int64).max:
                    df[col] = df[col].astype(np.int64)
            else:
                if c_min > np.finfo(np.float16).min and c_max < np.finfo(np.float16).max:
                    df[col] = df[col].astype(np.float16)
                elif c_min > np.finfo(np.float32).min and c_max < np.finfo(np.float32).max:
                    df[col] = df[col].astype(np.float32)
                else:
                    df[col] = df[col].astype(np.float64)
    end_mem = df.memory_usage().sum() / 1024**2
    if verbose: 
        print('Mem. usage decreased to {:5.2f} Mb ({:.1f}% reduction)'.format(end_mem, 100 * (start_mem - end_mem) / start_mem))
    return df

train=reduce_mem_usage(train)
test=reduce_mem_usage(test)
gc.collect()

train.to_csv('./dataset/train.csv')
test.to_csv('./dataset/test.csv')
