==> 4_Modeling.ipynb <==
import gc
import warnings
import numpy as np
import pandas as pd
import seaborn as sns
import matplotlib.pyplot as plt
from sklearn.preprocessing import StandardScaler

import xgboost as xgb
import lightgbm as lgb

warnings.filterwarnings('ignore')
pd.set_option('display.max_columns', None)

# Data Preprocessing

%%time
train = pd.read_csv('./dataset/train.csv')
test = pd.read_csv('./dataset/test.csv')

target = train['isFraud']

print(list(train.columns),"\n")
print(list(test.columns))

print(train.shape)
print(test.shape)

test_id = test['TransactionID']

train.drop(['Unnamed: 0', 'Unnamed: 0.1', 'Unnamed: 0.1.1', 'TransactionID'], axis=1, inplace=True)
test.drop(['Unnamed: 0', 'Unnamed: 0.1', 'Unnamed: 0.1.1', 'TransactionID'], axis=1, inplace=True)
print(train.shape)
print(test.shape)

train

train.drop(['isFraud'], axis=1, inplace=True)
print(train.shape)
print(test.shape)

# Label encoding provides poor result
from sklearn.preprocessing  import LabelEncoder

# One hot encoding our dataset
%%time
train = pd.get_dummies(train,drop_first=True)
test = pd.get_dummies(test,drop_first=True)

# XGBoost

%%time
# XGBOOST
xgmodel = xgb.XGBClassifier(colsample_bytree=0.4603, gamma=0.0468, 
                             learning_rate=0.05, max_depth=50, 
                             min_child_weight=1.7817, n_estimators=200,
                             reg_alpha=0.4640, reg_lambda=0.8571,
                             subsample=0.5213, random_state =7, nthread = -1)
xgmodel.fit(train, target)

y_pred = xgmodel.predict_proba(test)

sub = pd.DataFrame()
sub['TransactionID'] = test_id
sub['isFraud'] = y_pred[:, 1]
sub.to_csv('fesubmission.csv', index=False)

## Score: 0.915064

# LightGBM


clf = lgb.LGBMClassifier(
    num_leaves= 491,
          min_child_weight= 0.03454472573214212,
          feature_fraction= 0.3797454081646243,
          bagging_fraction= 0.4181193142567742,
          min_data_in_leaf= 106,
          objective= 'binary',
          max_depth= -1,
          learning_rate= 0.006883242363721497,
          boosting_type= "gbdt",
          bagging_seed= 11,
          metric= 'auc',
          verbosity= -1,
          reg_alpha= 0.3899927210061127,
          reg_lambda= 0.6485237330340494,
          random_state= 47,
)

clf.fit(train, target)

y_pred_lgbm = clf.predict_proba(test)

sub = pd.DataFrame()
sub['TransactionID'] = test_id
sub['isFraud'] = y_pred_lgbm[:, 1]
sub.to_csv('submission_lgbm.csv', index=False)

## Score: 0.896678
