==> 5_Ensemble.ipynb <==
from scipy.stats import spearmanr, rankdata
import pandas as pd
import numpy as np
import pickle

model1 = pd.read_csv("fesubmision.csv")
model2 = pd.read_csv("submission_lgbm.csv")
model1.head()

model1_col = model1['isFraud'].to_numpy()
model2_col = model2['isFraud'].to_numpy()

model1_col

ens = model1_col*2+model1_col
ens

sub = pd.DataFrame({
    'TransactionID' : model1['TransactionID'].values,
    'isFraud' : ens
})
sub.head()

sub.to_csv('Ensemble.csv', index=False)

- Result: 0.915064
- Didn't improve
